Allow filtering user search args in select2 search query
The ajax search query used to populate the meta box dropdown just gets
all users on the current site. On sites with a very high number of
users, this can lead to timeouts and database overloads. This commit
adds filters to the $user_args passed to the WP_User_Queryv search to
allow these queries to be filtered.