Fix PHP 7.4 notice in Admin_Ajax class
Adds a ! is_numeric check before checking if $val begins with 'u' in class Admin_Ajax. This fixes a 'Trying to access array offset on value of type int' notice in PHP 7.4 if an int would be passed in the $ignored array to get_possible_bylines_for_search.

This fix makes the unit tests pass on PHP 7.4